weeb_API/requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
djangorestframework==3.15.2
numpy==1.24.3
scikit-learn==1.3.0
django-cors-headers==4.3.1
